+++ torch_timeseries/experiments/iTransformer.py
@@ -5,7 +5,7 @@ import torch
 
 from torch_timeseries.dataloader.sliding_window_ts import SlidingWindowTS
 from torch_timeseries.utils.parse_type import parse_type
-from ..model import Informer
+from ..model import iTransformer
 from . import (
     ForecastExp,
     ImputationExp,
@@ -15,44 +15,39 @@ from . import (
 
 
 @dataclass
-class InformerParameters:
-    factor: int = 5
+class iTransformerParameters:
+    factor: int = 1
     d_model: int = 512
     n_heads: int = 8
     e_layers: int = 2
     d_layer: int = 1
-    d_ff: int = 512
-    dropout: float = 0.0
-    attn: str = "prob"
+    d_ff: int = 2048
+    dropout: float = 0.1
     embed: str = "timeF"
-    activation = "gelu"
-    distil: bool = True
-    mix: bool = True
+    activation:str= "gelu"
+    use_norm : bool = True
+    class_strategy : str = 'projection'
 
 @dataclass
-class InformerForecast(ForecastExp, InformerParameters):
-    model_type: str = "Informer"
+class iTransformerForecast(ForecastExp, iTransformerParameters):
+    model_type: str = "iTransformer"
         
     def _init_model(self):
         self.label_len = int(self.windows / 2)
         
-        self.model = Informer(
-            self.dataset.num_features,
-            self.dataset.num_features,
-            self.dataset.num_features,
-            self.pred_len,
+        self.model = iTransformer(
+            seq_len=self.windows,
+            pred_len=self.pred_len,
+            use_norm=self.use_norm,
+            class_strategy=self.class_strategy,
             factor=self.factor,
             freq=self.dataset.freq,
             d_model=self.d_model,
             n_heads=self.n_heads,
             e_layers=self.e_layers,
-            d_layers=self.d_layer,
             dropout=self.dropout,
-            attn=self.attn,
             embed=self.embed,
             activation=self.activation,
-            distil=self.distil,
-            mix=self.mix,
         )
         self.model = self.model.to(self.device)
 
@@ -83,54 +78,44 @@ class InformerForecast(ForecastExp, InformerParameters):
 
 
 
-class InformerClassModel(Informer):
+class iTransformerClassModel(iTransformer):
     def __init__(self, num_classes, **kwargs):
         super().__init__(**kwargs)
         self.num_classes = num_classes
-        self.t_projection = torch.nn.Linear(
-           int(  self.out_len/2), self.out_len)
 
         self.projection = torch.nn.Linear(
-           int( kwargs.get("d_model") * self.out_len), self.num_classes)
+           int( kwargs.get("d_model") * kwargs.get("c_in")), self.num_classes)
 
-    def forward(self, x_enc, x_mark_enc):
-        enc_out = self.enc_embedding(x_enc, None)
-        enc_out, attns = self.encoder(enc_out, attn_mask=None)
-        enc_out = self.t_projection(enc_out.transpose(1,2)).transpose(1,2)
+    def forward(self, x_enc):
+        _, T, N = x_enc.shape # B L N
         
-        # Output
-        # the output transformer encoder/decoder embeddings don't include non-linearity
+        enc_out = self.enc_embedding(x_enc, None) # covariates (e.g timestamp) can be also embedded as tokens
+        enc_out, attns = self.encoder(enc_out, attn_mask=None)
         output = torch.relu(enc_out)
-        # # zero-out padding embeddings
-        output = output * x_mark_enc.unsqueeze(-1)
-        # # (batch_size, seq_length * d_model)
         output = output.reshape(output.shape[0], -1)
         output = self.projection(torch.relu(output))
         return output
 
 @dataclass
-class InformerUEAClassification(UEAClassificationExp, InformerParameters):
-    model_type: str = "Informer"
+class iTransformerUEAClassification(UEAClassificationExp, iTransformerParameters):
+    model_type: str = "iTransformer"
 
     def _init_model(self):
-        self.label_len = self.windows/2
-        self.model = InformerClassModel(
-            enc_in=self.dataset.num_features,
-            dec_in=self.dataset.num_features,
-            c_out=self.dataset.num_features,
-            out_len=self.windows,
+        self.model = iTransformerClassModel(
+            seq_len=self.windows,
+            pred_len=self.windows,
+            use_norm=self.use_norm,
+            class_strategy=self.class_strategy,
             factor=self.factor,
+            freq='h',
             d_model=self.d_model,
             n_heads=self.n_heads,
             e_layers=self.e_layers,
             dropout=self.dropout,
-            attn=self.attn,
             embed=self.embed,
             activation=self.activation,
-            distil=self.distil,
-            mix=self.mix,
-            
-            num_classes=self.dataset.num_classes
+            num_classes=self.dataset.num_classes,
+            c_in=self.dataset.num_features
         )
         self.model = self.model.to(self.device)
 
@@ -146,47 +131,45 @@ class InformerUEAClassification(UEAClassificationExp, InformerParameters):
         batch_y = batch_y.to(self.device, dtype=torch.float32)
         padding_masks = padding_masks.to(self.device, dtype=torch.float32)
         
-        outputs = self.model(batch_x, padding_masks)  # torch.Size([batch_size, output_length, num_nodes])
+        outputs = self.model(batch_x)  # torch.Size([batch_size, output_length, num_nodes])
         return outputs, batch_y.long().squeeze(-1)
 
 
-class AnomalyModel(Informer):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.projection = torch.nn.Linear(
-            kwargs.get('d_model') , kwargs.get('c_out'))
-        self.t_projection = torch.nn.Linear(
-        int(self.out_len/2) ,self.out_len)
+# class AnomalyModel(iTransformer):
+#     def __init__(self, **kwargs):
+#         super().__init__(**kwargs)
+#         self.projection = torch.nn.Linear(
+#             kwargs.get('d_model') , kwargs.get('c_out'))
 
-    def forward(self, batch_x, batch_x_date_enc, dec_inp, dec_inp_date_enc):
-        enc_out = self.enc_embedding(batch_x, None)
-        enc_out, attns = self.encoder(enc_out, attn_mask=None)
-        # final
-        dec_out = self.projection(enc_out)
-        dec_out = self.t_projection(dec_out.transpose(1,2)).transpose(1,2)
-        return dec_out
+#     def forward(self, batch_x, batch_x_date_enc, dec_inp, dec_inp_date_enc):
+        
+#         B,T,N = batch_x.size()
+        
+#         enc_out = self.enc_embedding(batch_x, None)
+#         enc_out, attns = self.encoder(enc_out, attn_mask=None)
+#         # final
+#         dec_out = self.projection(enc_out)
+#         return dec_out[:, :, :N]
     
 @dataclass
-class InformerAnomalyDetection(AnomalyDetectionExp, InformerParameters):
-    model_type: str = "Informer"
+class iTransformerAnomalyDetection(AnomalyDetectionExp, iTransformerParameters):
+    model_type: str = "iTransformer"
 
     def _init_model(self):
         self.label_len = self.windows/2
-        self.model = AnomalyModel(
-            enc_in=self.dataset.num_features,
-            dec_in=self.dataset.num_features,
-            c_out=self.dataset.num_features,
-            out_len=self.windows,
+        self.model = iTransformer(
+            seq_len=self.windows,
+            pred_len=self.windows,
+            use_norm=self.use_norm,
+            class_strategy=self.class_strategy,
             factor=self.factor,
+            freq='h',
             d_model=self.d_model,
             n_heads=self.n_heads,
             e_layers=self.e_layers,
             dropout=self.dropout,
-            attn=self.attn,
             embed=self.embed,
             activation=self.activation,
-            distil=self.distil,
-            mix=self.mix,
         )
         self.model = self.model.to(self.device)
 
@@ -204,45 +187,39 @@ class InformerAnomalyDetection(AnomalyDetectionExp, InformerParameters):
 
 
 
-class ImputationModel(Informer):
+class ImputationModel(iTransformer):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.projection = torch.nn.Linear(
-            kwargs.get('d_model') , kwargs.get('c_out'))
-        self.t_projection = torch.nn.Linear(
-        int(self.out_len/2) ,self.out_len)
+            kwargs.get('d_model') , kwargs.get('pred_len'))
 
     def forward(self, batch_x, batch_x_date_enc, dec_inp, dec_inp_date_enc):
         enc_out = self.enc_embedding(batch_x, batch_x_date_enc)
         enc_out, attns = self.encoder(enc_out, attn_mask=None)
         # final
         dec_out = self.projection(enc_out)
-        dec_out = self.t_projection(dec_out.transpose(1,2)).transpose(1,2)
         return dec_out
     
     
 @dataclass
-class InformerImputation(ImputationExp, InformerParameters):
-    model_type: str = "Informer"
+class iTransformerImputation(ImputationExp, iTransformerParameters):
+    model_type: str = "iTransformer"
 
     def _init_model(self):
         self.label_len = self.windows/2
-        self.model = ImputationModel(
-            enc_in=self.dataset.num_features,
-            dec_in=self.dataset.num_features,
-            c_out=self.dataset.num_features,
-            out_len=self.windows,
+        self.model = iTransformer(
+            seq_len=self.windows,
+            pred_len=self.windows,
+            use_norm=self.use_norm,
+            class_strategy=self.class_strategy,
             factor=self.factor,
             freq=self.dataset.freq,
             d_model=self.d_model,
             n_heads=self.n_heads,
             e_layers=self.e_layers,
             dropout=self.dropout,
-            attn=self.attn,
             embed=self.embed,
             activation=self.activation,
-            distil=self.distil,
-            mix=self.mix,
         )
         self.model = self.model.to(self.device)