+++ torch_timeseries/model/__init__.py
@@ -3,3 +3,4 @@ from .Autoformer import Autoformer
 from .FEDformer import FEDformer
 from .Informer import Informer
 from .PatchTST import PatchTST
+from .iTransformer import iTransformer