+++ CHANGELOG.md
@@ -6,3 +6,19 @@ fixed: dependencies issue of torch-timeserie pypi package
 
 func: Forecast/Imputation/UEAClassification AnomalyDetection experiments
 func: pytexp entrypoint
+
+
+## 0.1.1
+fix: fix a typo of the standard scaler
+fix: fix some cli errors
+
+## 0.1.2
+func: add Informer Forecast/Imputation/AnomalyDetection/UEAClassification
+func: add Autoformer Forecast/Imputation/AnomalyDetection/UEAClassification
+func: add FEDformer Forecast/Imputation/AnomalyDetection/UEAClassification
+func: add PatchTST Forecast/Imputation/AnomalyDetection/UEAClassification
+
+## 0.1.3
+
+func: add iTransformer Forecast/Imputation/AnomalyDetection/UEAClassification
+fix: fix a typo of Informer d_layer config